api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg[binary]==3.2.6
python-dotenv==1.0.0
requests==2.31.0
cloudinary==1.36.0
google-generativeai==0.3.1
google-search-results==2.4.2
Pillow==10.4.0
gunicorn==21.2.0
